# openapi.yaml
openapi: 3.0.3
info:
  title: ApiMe
  version: "1.3.0"
  description: |
    API para orquestração de múltiplas instâncias de WhatsApp.

    Autenticação: header `Authorization: Bearer <token>`, com três tipos de credencial
    (userJwt, apiToken, instanceToken). O middleware valida em cascata e para no primeiro
    que reconhecer o token.

    Escopo: `instanceToken` só alcança a própria instância. As rotas de perfil
    (`/instances/{id}/profile/*`, `/instances/{id}/business/*`) exigem `instanceToken` e
    recusam os demais. As rotas `/users/*` exigem usuário com role admin.
servers:
  - url: http://localhost:8080/api
    description: Local
paths:
...
  /instances/{id}/messages/text:
    post:
      summary: Enviar texto
      tags: [Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, text]
              properties:
                to:
                  type: string
                  example: "5511999999999"
                text:
                  type: string
                quoted:
                  type: string
                  description: Message ID (HEX) da mensagem citada
      responses:
        "409":
          description: Requisição com a mesma Idempotency-Key ainda em andamento
        "503":
          description: Sessão não pronta. Nada foi enviado e a Idempotency-Key é liberada para nova tentativa
        "422":
          description: Idempotency-Key já usada com outro conteúdo
        "200":
          description: Enviado
  /instances/{id}/messages/media:
    post:
      summary: Enviar mídia (imagem, vídeo, GIF ou figurinha)
      tags: [Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [to, type, file]
              properties:
                to:
                  type: string
                  description: JID do destinatário
                type:
                  type: string
                  enum: [image, video, gif, sticker]
                  description: >-
                    Tipo de mídia. "gif" é enviado como vídeo MP4 com reprodução
                    em laço (o WhatsApp não aceita arquivo .gif: ele chegaria
                    como imagem parada). "sticker" exige WebP 512x512 de no
                    máximo 500 KB, animado ou não.
                file:
                  type: string
                  format: binary
                  description: >-
                    Arquivo. Imagem e vídeo aceitam os formatos usuais; "gif"
                    espera MP4; "sticker" espera WebP 512x512 até 500 KB.
                caption:
                  type: string
                  description: >-
                    Legenda (opcional). Ignorada em "sticker": o protocolo não
                    tem esse campo na figurinha.
                quoted:
                  type: string
                  description: ID da mensagem citada (opcional)
      responses:
        "409":
          description: Requisição com a mesma Idempotency-Key ainda em andamento
        "503":
          description: Sessão não pronta. Nada foi enviado e a Idempotency-Key é liberada para nova tentativa
        "422":
          description: Idempotency-Key já usada com outro conteúdo
        "200":
          description: Enviado
  /instances/{id}/messages/audio:
    post:
      summary: Enviar áudio
      tags: [Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [to, file]
              properties:
                to:
                  type: string
                  description: JID do destinatário
                file:
                  type: string
                  format: binary
                  description: Arquivo de áudio
                ptt:
                  type: boolean
                  description: Push-to-Talk (áudio de voz)
                  default: false
                seconds:
                  type: integer
                  description: Duração em segundos
                quoted:
                  type: string
                  description: ID da mensagem citada (opcional)
      responses:
        "409":
          description: Requisição com a mesma Idempotency-Key ainda em andamento
        "503":
          description: Sessão não pronta. Nada foi enviado e a Idempotency-Key é liberada para nova tentativa
        "422":
          description: Idempotency-Key já usada com outro conteúdo
        "200":
          description: Enviado
  /instances/{id}/messages/document:
    post:
      summary: Enviar documento
      tags: [Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [to, file]
              properties:
                to:
                  type: string
                  description: JID do destinatário
                file:
                  type: string
                  format: binary
                  description: Arquivo do documento
                caption:
                  type: string
                  description: Legenda (opcional)
                filename:
                  type: string
                  description: Nome do arquivo (opcional)
                quoted:
                  type: string
                  description: ID da mensagem citada (opcional)
      responses:
        "409":
          description: Requisição com a mesma Idempotency-Key ainda em andamento
        "503":
          description: Sessão não pronta. Nada foi enviado e a Idempotency-Key é liberada para nova tentativa
        "422":
          description: Idempotency-Key já usada com outro conteúdo
        "200":
          description: Enviado
...
  /instances/{id}/messages:
    post:
      summary: Enfileirar mensagem genérica
      tags: [Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, type, payload]
              properties:
                to:
                  type: string
                type:
                  type: string
                  description: Tipo da mensagem
                payload:
                  type: string
                  description: Conteúdo serializado
      responses:
        "409":
          description: Requisição com a mesma Idempotency-Key ainda em andamento
        "503":
          description: Sessão não pronta. Nada foi enviado e a Idempotency-Key é liberada para nova tentativa
        "422":
          description: Idempotency-Key já usada com outro conteúdo
        "200":
          description: Mensagem enfileirada
    get:
      summary: Listar mensagens da instância
      tags: [Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
      responses:
        "200":
          description: Lista de mensagens
  /instances/{id}/messages/contact:
    post:
      summary: Enviar contato
      tags: [Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, displayName]
              properties:
                to:
                  type: string
                displayName:
                  type: string
                vcard:
                  type: string
                quoted:
                  type: string
                  description: ID da mensagem citada
                quotedParticipant:
                  type: string
                quotedText:
                  type: string
                quotedFromMe:
                  type: boolean
                mentionedJids:
                  type: array
                  items:
                    type: string
      responses:
        "409":
          description: Requisição com a mesma Idempotency-Key ainda em andamento
        "503":
          description: Sessão não pronta. Nada foi enviado e a Idempotency-Key é liberada para nova tentativa
        "422":
          description: Idempotency-Key já usada com outro conteúdo
        "200":
          description: Contato enviado
  /instances/{id}/messages/location:
    post:
      summary: Enviar localização
      tags: [Mensagens]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, latitude, longitude]
              properties:
                to:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
                name:
                  type: string
                address:
                  type: string
                quoted:
                  type: string
                quotedParticipant:
                  type: string
                quotedText:
                  type: string
                quotedFromMe:
                  type: boolean
      responses:
        "409":
          description: Requisição com a mesma Idempotency-Key ainda em andamento
        "503":
          description: Sessão não pronta. Nada foi enviado e a Idempotency-Key é liberada para nova tentativa
        "422":
          description: Idempotency-Key já usada com outro conteúdo
        "200":
          description: Localização enviada
...
  /instances/{id}/stories/text:
    post:
      summary: Publicar story de texto
      tags: [Stories]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - $ref: "#/components/parameters/idempotencyKey"
      description: >-
        Publica um story de texto. A audiência não é informada pelo chamador:
        vale a configuração de privacidade de status da conta, consultável em
        GET /instances/{id}/whatsapp/status-privacy. Não confundir com
        POST /instances/{id}/whatsapp/status, que define o recado do perfil.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                backgroundArgb:
                  type: integer
                  format: int64
                  description: >-
                    Cor de fundo do card, em ARGB opaco (ex: 4280427042 para 0xFF25D366).
                    Omitido usa o verde do WhatsApp. Um story de texto sempre carrega fundo,
                    então a apime define um de qualquer forma.
                textArgb:
                  type: integer
                  format: int64
                  description: Cor do texto, em ARGB opaco. Omitido usa branco.
                font:
                  type: string
                  enum: [system, system_text, system_bold, fb_script, morningbreeze_regular, calistoga_regular, exo2_extrabold, courierprime_bold]
                  description: Fonte do card. Valor desconhecido cai na fonte do sistema.
      responses:
        "409":
          description: Requisição com a mesma Idempotency-Key ainda em andamento
        "503":
          description: Sessão não pronta. Nada foi enviado e a Idempotency-Key é liberada para nova tentativa
        "422":
          description: Idempotency-Key já usada com outro conteúdo
        "200":
          description: Publicado
        "403":
          description: Token não é da instância, ou não é token de instância
  /instances/{id}/stories/media:
    post:
      summary: Publicar story de imagem ou vídeo
      tags: [Stories]
      security: [{instanceToken: []}]
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - $ref: "#/components/parameters/idempotencyKey"
      description: >-
        Publica um story de imagem ou vídeo. A audiência não é informada pelo
        chamador: vale a configuração de privacidade de status da conta,
        consultável em GET /instances/{id}/whatsapp/status-privacy.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [type, file]
              properties:
                type:
                  type: string
                  enum: [image, video]
                file:
                  type: string
                  format: binary
                caption:
                  type: string
                  description: Legenda (opcional)
      responses:
        "409":
          description: Requisição com a mesma Idempotency-Key ainda em andamento
        "503":
          description: Sessão não pronta. Nada foi enviado e a Idempotency-Key é liberada para nova tentativa
        "422":
          description: Idempotency-Key já usada com outro conteúdo
        "200":
          description: Publicado
        "403":
          description: Token não é da instância, ou não é token de instância
components:
  securitySchemes:
    userJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT de sessão, obtido em POST /auth/login. Expira conforme JWT_EXP_HOURS.
        Identifica um usuário (claim sub).
    apiToken:
      type: http
      scheme: bearer
      description: >
        Token de integração, gerado por um admin em POST /users/{id}/token. Não expira.
        Identifica o mesmo usuário do JWT, para uso servidor a servidor.
    instanceToken:
      type: http
      scheme: bearer
      description: >
        Token com escopo de UMA instância, entregue na criação e trocável em
        POST /instances/{id}/token/rotate. Só acessa a própria instância.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Depreciado, mantido para não quebrar quem já referencia este nome.
        Use userJwt, apiToken ou instanceToken conforme o caso.
  parameters:
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 255
      description: >
        Chave gerada pelo cliente (UUID v4 serve) que torna seguro repetir o envio.
        A primeira chamada guarda o resultado por 24h, e uma repetição com a mesma
        chave devolve o mesmo resultado, com o header X-Idempotent-Replay: true,
        em vez de mandar a mensagem de novo.
    instanceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...